+++ mss/write-cross-frame-by-index.cpp
@@ -6,7 +6,7 @@
 void FORCE_NOINLINE helper(const char* b) {
   long long delta = 0;
   charBuffer dummy;
-  GET_DISTANCE(delta, b, dummy.data);
+  GET_DISTANCE(delta, (long long)b, (long long)dummy.data);
   update_by_index(dummy, delta, 8, 1, 'c');
 }
